Move matchPatterns
