+++ src/matchPattern.ts
