+++ tsconfig.json
@@ -1,12 +1,22 @@
 {
   "compilerOptions": {
     "paths": {
-      "~roll": ["./src/roll/index.ts"],
-      "~Die": ["./src/Die/index.ts"],
-      "~constants": ["./src/constants.ts"],
-      "~types": ["./src/types.ts"]
+      "~roll": [
+        "./src/roll/index.ts"
+      ],
+      "~Die": [
+        "./src/Die/index.ts"
+      ],
+      "~matchPattern": [
+        "./src/matchPattern.ts"
+      ],
+      "~types": [
+        "./src/types.ts"
+      ]
     },
-    "lib": ["ESNext"],
+    "lib": [
+      "ESNext"
+    ],
     "module": "esnext",
     "target": "esnext",
     "moduleResolution": "bundler",
@@ -20,6 +30,8 @@
     "allowSyntheticDefaultImports": true,
     "forceConsistentCasingInFileNames": true,
     "allowJs": false,
-    "types": ["bun-types"]
+    "types": [
+      "bun-types"
+    ]
   }
 }
